global_wellness_auth/requirements.txt
Flask==3.0.0
# Flask-SQLAlchemy 3.0.5 is compatible with SQLAlchemy 1.4.x (required by Rasa)
Flask-SQLAlchemy==3.0.5
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
pytest==7.4.3
Werkzeug==3.0.1
# SQLAlchemy version must be compatible with Rasa (Rasa requires <1.5.0)
SQLAlchemy>=1.4.0,<1.5.0
rasa>=3.6.0,<3.7.0
rasa-sdk>=3.6.0,<3.7.0
requests==2.31.0

